+++ hexrd/legacy_transforms/debug_helpers.h
